[Fix] Remember the correct MIDI input device when MIDI device IDs get reordered (https://bugs.openmpt.org/view.php?id=756).
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@6093 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Mpdlgs.cpp
@@ -1826,7 +1826,7 @@ void CMidiSetupDlg::OnOK()
 	if ((combo = (CComboBox *)GetDlgItem(IDC_COMBO1)) != NULL)
 	{
 		int n = combo->GetCurSel();
-		if (n >= 0) m_nMidiDevice = static_cast<LONG>(combo->GetItemData(n));
+		if (n >= 0) m_nMidiDevice = static_cast<UINT>(combo->GetItemData(n));
 	}
 
 	TrackerSettings::Instance().aftertouchBehaviour = static_cast<RecordAftertouchOptions>(m_ATBehaviour.GetItemData(m_ATBehaviour.GetCurSel()));
